% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{aggregate_all_permutations}
\alias{aggregate_all_permutations}
\title{Aggregate across selected categories of quantity flow data, looping through all permutations}
\usage{
aggregate_all_permutations(input_data,
  collapse_vars = DS_IDVARS_WITH_ALLS, ...)
}
\arguments{
\item{input_data}{data table with quantity (flow) variables}

\item{collapse_vars}{character string indicating the variables (columns) whose values will be collapsed (aggregated)
to "All"}
}
\description{
Inputs are a data frame with model data, and a character vector of ID variables to loop through. All possible
permutations of the variables provided are returned in the output, with no redundant categories written out.
Returns the intial data frame with permutations appended.
}
\details{
The method implemented here assumes that all itemized elements of each variable to add to the total. Any
  excluded elements will not be part of the reported total, and redundant categories (e.g., gasoline,
  total liquid fuels) will be double counted.
}
