% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{aggregate_item_regions}
\alias{aggregate_item_regions}
\title{Aggregate country-level quantity flow data to iTEM analysis regions}
\usage{
aggregate_item_regions(downscaled_data,
  item_region_map = load_country_region_mappings()[["item"]],
  compute_global_totals = FALSE, ...)
}
\arguments{
\item{downscaled_data}{data frame with downscaled (i.e. country-level) data. Can also include global totals, assigned
to iso code "All".}

\item{item_region_map}{data table mapping from 3-digit ISO code to iTEM region, called from
(\code{\link{load_country_region_mappings}})}

\item{compute_global_totals}{logical indicating whether to compute the global totals. The default is FALSE, as it is
assumed that this step has been conducted under (\code{\link{aggregate_all_permutations}}).}
}
\description{
This function takes in a data frame with quantity flow data by ISO code and a mapping from ISO code to iTEM region.
It joins in the appropriate iTEM region for each country ISO code, aggregates by these regions, and returns a
dataframe.
}
\details{
The mapping from ISO code to iTEM region is user-specified with a provided default.
}
