% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{approx_fun}
\alias{approx_fun}
\title{approx_fun}
\usage{
approx_fun(year, value, rule = 1)
}
\arguments{
\item{year}{Numeric year, in a melted (gathered) data frame}

\item{value}{Numeric value to interpolate}

\item{rule}{Rule to use; see \code{\link{approx}} and details}
}
\value{
Interpolated values.
}
\description{
\code{\link{approx}} (interpolation) for interpolating years in a dplyr pipeline.
}
\details{
Rule=1: interpolate only; out of bounds -> missing values
Rule=2: out of bounds -> fixed extrapolation
}
\examples{
df <- data.frame(year = 1:5, value = c(1, 2, NA, 4, 5))
approx_fun(df$year, df$value, rule = 2)
}
