% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{assign_socioeconomic_scenario}
\alias{assign_socioeconomic_scenario}
\title{Assign country-level socioeconomic scenario to each model/scenario}
\usage{
assign_socioeconomic_scenario(data_list, method = "exogenous",
  SSE_variable = "PPP-GDP")
}
\arguments{
\item{data_list}{list of data tables wherein each element is a model-specific dataframe}

\item{method}{select a method for assigning scenarios.
SSE: minimize sum of squared errors.
exogenous: use a default socioeconomic scenario for downscaling all models and scenarios. can be
over-ridden for specific model x scenario in exogenous mapping files.}

\item{variable}{variable to use in determining the assignments; can be population or gdp}
}
\description{
Assign country-level socioeconomic scenario to each model/scenario
}
\details{
Returns a list of dataframes where each model is an element. Data frames assign model's scenarios
to a country-level socioeconomics file.
}
