% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_elasticities.R
\name{compute_elasticities}
\alias{compute_elasticities}
\title{Compute elasticities from downscaled iTEM dataset}
\usage{
compute_elasticities(item_data, number_variable = "pkm_percapita",
  base_variable = "PPP-GDP_percapita", year_range = c(2010, 2050),
  models = "All", regions = "Global", return_boxplot = TRUE, ...)
}
\arguments{
\item{number_variable}{number of the logarithm}

\item{base_variable}{base used in the logarithm}

\item{year_range}{timespan to use in computing elasticities.}

\item{models}{which models to use (default = All)}

\item{regions}{which regions to use (default = Global)}

\item{data}{data table with the variables necessary to perform the computations}
}
\description{
Compute elasticities from downscaled iTEM dataset
}
\details{
This function computes and produces boxplots of the elasticities between specified variables of interest
}
