% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{compute_global_totals}
\alias{compute_global_totals}
\title{Compute global totals for quantity (flow) data}
\usage{
compute_global_totals(input_data, region_col = "region",
  append_to_df = TRUE)
}
\arguments{
\item{input_data}{data table with quantity (flow) variables}

\item{region_col}{string indicating the name of the region column that is to be aggregated}

\item{append_to_df}{logical (default = TRUE) indicating whether to return a data table with the global totals
appended to the initial data table (TRUE), or whether to only return the global totals (FALSE)}
}
\description{
Helper function to compute global totals from either country-level or region-level flow (i.e. quantity) data
}
\details{
The method implemented here assumes that all itemized countries or regions add to a global total. Any
  excluded countries/regions will not be part of the reported global total, and redundant regions (e.g., OECD,
  Europe) will be double counted.
}
