% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_bp.R
\name{correct_bp}
\alias{correct_bp}
\title{Apply corrections to BP submitted data}
\usage{
correct_bp(x)
}
\arguments{
\item{x}{data table with BP model data}
}
\description{
Apply corrections to BP submitted data
}
\details{
Applies model-specific corrections to reporting categories
}
