% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_eia.R
\name{correct_eia}
\alias{correct_eia}
\title{Apply corrections to EIA (WEPS+) submitted data}
\usage{
correct_eia(x)
}
\arguments{
\item{x}{data table with EIA (WEPS+) model data}
}
\description{
Apply corrections to EIA (WEPS+) submitted data
}
\details{
Applies model-specific corrections to reporting categories
}
