% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_get.R
\name{correct_get}
\alias{correct_get}
\title{Apply corrections to GET submitted data}
\usage{
correct_get(x)
}
\arguments{
\item{x}{data table with GET model data}
}
\description{
Apply corrections to GET submitted data
}
\details{
Applies model-specific corrections to reporting categories
}
