% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_itf.R
\name{correct_itf}
\alias{correct_itf}
\title{Apply corrections to ITF submitted data}
\usage{
correct_itf(x)
}
\arguments{
\item{x}{data table with ITF model data}
}
\description{
Apply corrections to ITF submitted data
}
\details{
Applies model-specific corrections to reporting categories
}
