% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_message.R
\name{correct_message}
\alias{correct_message}
\title{Apply corrections to MESSAGE submitted data}
\usage{
correct_message(x)
}
\arguments{
\item{x}{data table with MESSAGE model data}
}
\description{
Apply corrections to MESSAGE submitted data
}
\details{
Applies model-specific corrections to reporting categories
}
