% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_momo.R
\name{correct_momo}
\alias{correct_momo}
\title{Apply corrections to MoMo submitted data}
\usage{
correct_momo(x)
}
\arguments{
\item{x}{data table with MoMo model data}
}
\description{
Apply corrections to MoMo submitted data
}
\details{
Applies model-specific corrections to reporting categories
}
