% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{correct_preprocessed_data}
\alias{correct_preprocessed_data}
\title{Apply corrections to pre-processed data, where the submitted data requires further modifications}
\usage{
correct_preprocessed_data(data)
}
\arguments{
\item{data}{list of data frames with model data}
}
\description{
Apply corrections to pre-processed data, where the submitted data requires further modifications
}
\details{
Applies model-specific functions, matching the named element of the list to a
corresponding function containing the same string ("correct_" + model name)
}
