% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_roadmap.R
\name{correct_roadmap}
\alias{correct_roadmap}
\title{Apply corrections to Roadmap submitted data}
\usage{
correct_roadmap(x)
}
\arguments{
\item{x}{data table with Roadmap model data}
}
\description{
Apply corrections to Roadmap submitted data
}
\details{
Applies model-specific corrections to reporting categories
}
