% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_statoil.R
\name{correct_statoil}
\alias{correct_statoil}
\title{Apply corrections to Statoil submitted data}
\usage{
correct_statoil(x)
}
\arguments{
\item{x}{data table with Statoil model data}
}
\description{
Apply corrections to Statoil submitted data
}
\details{
Applies model-specific corrections to reporting categories
}
