% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{derive_variables}
\alias{derive_variables}
\title{Derive new variables from existing variables in a dataset}
\usage{
derive_variables(model_data_df,
  mapping = get_variable_mapping("downscale/indicators.csv"),
  bind_derived_data = TRUE, ...)
}
\arguments{
\item{model_data_df}{data frame with pre-processed/corrected iTEM data in the appropriate format}

\item{mapping}{dataframe specifying the variables to be created, along with instructions
for creating them and the units of the new variable}

\item{bind_derived_data}{logical (default = TRUE) indicating whether to return a data frame with the original model
data bound to the output of this function (TRUE), or to only return the variables calculated by this function
(FALSE)}
}
\description{
This function takes in a dataframe with model data, and a variable derivation CSV table with
instructions detailing what the new variables will be named, how they will be calculated, and what
their units will be.
}
\details{
This function derives new variables from existing variables in the data frame. Each new variable is derived
  from two and only two existing variables. The operations allowed are +, *, and /. Unit conversions are applied
  multiplicatively. If the operation is addition and there is a unit conversion provided, the conversion is applied
  to the second variable in the derivation.
}
