% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{extract_global_data_to_df}
\alias{extract_global_data_to_df}
\title{Extract global data from model-specific data list to a data frame.}
\usage{
extract_global_data_to_df(model_data, input_region_col = "region",
  output_region_col = "region", ...)
}
\arguments{
\item{model_data}{list of model-specific data frames from which to extract the global data}

\item{input_region_col}{name of the region column in the input data}

\item{output_region_col}{name of the region column in the output (return) data}
}
\description{
Simple wrapper function to extract global data (region == "All") from list of model output data. Returns global-only
data if executed after (\code{\link{remove_redundant_alls}}), as called from
(\code{\link{prepare_preprocessed_data}})
}
\details{
Modeling teams may provide some data at the global level only, i.e., with no regional detail available. Such
  data is neither downscaled to countries/regions nor dropped. This function generates a data frame with global-only
  data from all models that can be binded to other data frames with regional detail.
}
