% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{find_data_file}
\alias{find_data_file}
\title{find_data_file}
\usage{
find_data_file(filename, optional = FALSE, quiet = FALSE)
}
\arguments{
\item{filename}{Filename (extension optional) to find}

\item{optional}{Logical: file optional to find?}

\item{quiet}{Logical - suppress messages?}
}
\value{
Fully qualified file name, or NULL if file not found but is optional.
}
\description{
Find an internal, i.e. included with the package, data file.
}
\details{
Throws an error if file not found (and file is not optional).
}
