% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{generate_ds_proxy_data}
\alias{generate_ds_proxy_data}
\title{Generate downscaling proxy data for scenarios, countries, iTEM years, and iTEM variable classes}
\usage{
generate_ds_proxy_data(pop_data_fn = "downscale/SSP_Population.csv",
  gdp_data_fn = "downscale/SSP_GDP_PPP.csv",
  co2_data_t0_fn = "downscale/CDIAC_CO2_ctry.csv",
  transportenergy_data_t0_fn = "downscale/IEA_trn_ctry.csv",
  use_modal_detail = TRUE, apply_modal_elasticities = TRUE,
  save_output = TRUE, create_dir = FALSE, output_dir = DS_OUTPUT_DIR, ...)
}
\arguments{
\item{pop_data_fn}{file name of data with population by scenario, country, year (including future iTEM years)}

\item{gdp_data_fn}{file name of data with GDP by scenario, country, year (including future iTEM years)}

\item{co2_data_t0_fn}{file name of data with CO2 emissions by country, for a base year}

\item{transportenergy_data_t0_fn}{file name of data with transportation energy consumption data by country,
for a base year. May include mode-level detail.}

\item{use_modal_detail}{logical indicating whether to determine mode-specific proxy values. If false,
a single proxy will be calculated for the whole transportation sector.}

\item{apply_modal_elasticities}{logical indicating whether to differentiate income elasticities by mode.
If true, exogenous assumptions are used. If false, all modes are assigned an income elasticity of 1.}

\item{save_output}{indicate whether to save the output of this function}

\item{create_dir}{indicate whether to create a new directory if the specified one does not already exist}

\item{output_dir}{file path where output of this function is saved}
}
\description{
Generate downscaling proxy data for scenarios, countries, iTEM years, and iTEM variable classes
}
\details{
Derives country-level downscaling proxy data for all iTEM years. Does not apply or determine
model-specific growth elasticities.
}
