% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{get_country_region_map}
\alias{get_country_region_map}
\title{Get the country-to-region mapping for a given model}
\usage{
get_country_region_map(model_name)
}
\arguments{
\item{model_name}{name of model whose country-to-region mapping is to be returned}
}
\description{
Get the country-to-region mapping for a given model
}
\details{
Takes in yaml lists of countries in each model region, and returns a data frame with two columns: iso and
  region
}
