% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{load_country_region_mappings}
\alias{load_country_region_mappings}
\title{Get the country-to-region mappings for the project and optionally for a set of models}
\usage{
load_country_region_mappings(model_data_folder = NA,
  item_map_name = "item", ...)
}
\arguments{
\item{model_data_folder}{folder where the pre-processed model output data is located. If NA, this will return a
list with a single internal (included with the package) country-to-region mapping table for iTEM project regions.}

\item{item_map_name}{the name of the dataframe with the iTEM project country-to-region mapping}
}
\description{
Returns a list of data frames, each named according to the data source/model, each of which has two columns:
iso and region.
}
\details{
This function is designed to return country-to-region mapping assignments as required by other functions in
  the package. If a model data folder is provided, the function requires that the mapping lists be within this folder
  as model/regions.yaml. If no such folder is provided, then the returned list will only have the internal
  (i.e., this project's) country-to-region mapping.
}
