% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{load_data_file}
\alias{load_data_file}
\title{load_data_file}
\usage{
load_data_file(filename, optional = FALSE, quiet = FALSE)
}
\arguments{
\item{filename}{filename of file to load}

\item{optional}{Logical indicating whether file is optional}

\item{quiet}{Logical - suppress messages?}
}
\value{
A data frame (csv) or yaml list.
}
\description{
Load an internal, i.e. included with the package, data file.
}
\details{
Optional files that are not found are returned as NA.
}
