% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-.R
\name{load_model_data}
\alias{load_model_data}
\title{Load the iTEM model database}
\usage{
load_model_data(version)
}
\arguments{
\item{version}{Model database version to load (1 or 2)}
}
\description{
Load the iTEM model database
}
