% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_roadmap.R
\name{load_roadmap_tech_mapping}
\alias{load_roadmap_tech_mapping}
\title{Load Roadmap technology re-mapping CSV table}
\usage{

  load_roadmap_tech_mapping(roadmap_remapping_folder = "roadmap model processed",
  filename = "tech_mapping.csv")
}
\arguments{
\item{roadmap_remapping_folder}{Folder where Roadmap remapping file is located}

\item{filename}{Name of Roadmap tech remapping file}
}
\description{
Load Roadmap technology re-mapping CSV table
}
