% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{perform_item_data_processing}
\alias{perform_item_data_processing}
\title{Perform iTEM data processing, producing region- and/or country-level database(s) from raw model output}
\usage{
perform_item_data_processing(model_data_folder, output_folder,
  write_item_region_data = TRUE, write_item_country_data = FALSE,
  return_output = FALSE, spread_by_years = TRUE, ...)
}
\arguments{
\item{model_data_folder}{folder where the pre-processed model output data is located}

\item{output_folder}{folder where the harmonized output dataset will be saved}

\item{write_item_region_data}{logical (default = TRUE) indicating whether to write out a dataset aggregated to item
reporting regions. If FALSE, no iTEM region level data will be returned.}

\item{write_item_country_data}{logical (default = FALSE) indicating whether to write out a dataset with all variables
reported at the country level. Warning: the dataset may be quite large.}

\item{return_output}{logical (default = FALSE) indicating whether to return an object in the R environment}

\item{spread_by_years}{logical (default = TRUE) indicating whether to "spread" the output so that the reporting years
are columns and values are listed within the year columns}
}
\description{
This is a wrapper that performs all steps for inter-model harmonization in iTEM: data loading, pre-processing,
determination of nation-level socioeconomic scenarios for downscaling, downscaling, computation of "All" subtotals,
re-aggregation (if requested), deriving new variables from the existing set, and saving the output.
}
\details{
This function applies nearly all other item downscaling functions. While it only requires a few arguments,
  any additional arguments provided will be passed to other functions. The list of functions called is provided here:
  \code{\link{load_preprocessed_data}}, \code{\link{prepare_preprocessed_data}},
  \code{\link{load_country_region_mappings}}, \code{\link{compute_country_shares}},
  \code{\link{assign_socioeconomic_scenario}}, \code{\link{downscale_flow_variables}},
  \code{\link{extract_global_data_to_df}}, \code{\link{aggregate_all_permutations}},
  \code{\link{aggregate_item_regions}}, \code{\link{derive_variables}}, \code{\link{save_output}}
}
