% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{prepare_preprocessed_data}
\alias{prepare_preprocessed_data}
\title{Apply corrections to pre-processed data}
\usage{
prepare_preprocessed_data(model_data_list, subset_quantity_flows = TRUE,
  remove_redundant_alls = TRUE, interpolate_years = TRUE,
  drop_na_values = TRUE, aggregate_redundant_categories = TRUE, ...)
}
\arguments{
\item{model_data_list}{list of data frames with model-specific data}

\item{subset_quantity_flows}{logical (default = TRUE) indicating whether to subset only quantity flow variables}

\item{remove_redundant_alls}{logical (default = TRUE) indicating whether to remove observations with "All" that could
be calculated instead by adding provided components. These "All" values should not be downscaled.}

\item{interpolate_years}{logical (default = TRUE) indicating whether to interpolate years.}

\item{drop_na_values}{logical (default = TRUE) indicating whether to drop observations with no value in model-
submitted data. If FALSE, missing values are set to zero.}

\item{aggregate_redundant_categories}{logical (default = TRUE) indicating whether to add up categories that are
redundant (i.e., all ID information the same), which generally result from re-setting variables from model-
reported data to more aggregate categories.}
}
\description{
Addresses errors and omissions in submitted model data loaded by \code{\link{load_preprocessed_data}} prior to
performing any data processing. Where necessary, corrections are performed by calling model-specific functions. This
function also linearly interpolates submitted data in order to standardize iTEM analysis years.
}
\details{
Applies model-specific functions, searching for correction functions for each model name in the list (i.e.,
  "correct_" + model name). The interpolation function does not extrapolate, so the data from any model does not
  necessarily include all iTEM analysis years.
}
