% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{prepare_transportenergy_t0_data}
\alias{prepare_transportenergy_t0_data}
\title{Generate base-year transportation dataset from which to construct the downscaling proxy}
\usage{
prepare_transportenergy_t0_data(country_data, use_modal_detail = TRUE,
  mode_mapping_fn = "downscale/map_mode_IEA.csv", ...)
}
\arguments{
\item{datafile}{data table with transportation energy consumption}

\item{mode_mapping_file}{mapping from mode in the transportation energy data to the iTEM modes}
}
\description{
Generate base-year transportation dataset from which to construct the downscaling proxy
}
\details{
Derives country-level transportation energy consumption in the base-year by iTEM mode that will
be used for constructing the downscaling proxy dataset for transportation-related variables
}
