% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{repeat_add_columns}
\alias{repeat_add_columns}
\title{repeat_add_columns}
\usage{
repeat_add_columns(x, y)
}
\arguments{
\item{x}{Data frame to repeat}

\item{y}{A copy of \code{x} is created for each row of this data frame}
}
\value{
A repeated \code{x} with columns from \code{y} added.
}
\description{
Repeat a data frame for each entry in a second, binding the columns together.
}
\details{
This function is used to repeat a table by specified values of a variable, written to a new column as
  necessary.
}
\examples{
x <- tibble::tibble(x = 1:3)
y <- tibble::tibble(y = c(4, 5), z = c(6, 7))
repeat_add_columns(x, y)
}
