% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{assign_socioeconomic_scenario}
\alias{assign_socioeconomic_scenario}
\title{Assign country-level socioeconomic realization to each model+scenario}
\usage{
assign_socioeconomic_scenario(model_data_list, country_region_map_list,
  method = "SSE", SSE_variable = "PPP-GDP")
}
\arguments{
\item{model_data_list}{list of data tables of model-specific pre-processed, prepared data}

\item{country_region_map_list}{list of data tables of model-specific country-to-region mappings}

\item{method}{method used for assigning scenarios. SSE: default behavior; minimize sum of squared errors in future
ratios of variable indicated, by year and model region. exogenous: use exogenously provided assignments between
model+scenario and nation-level socioeconomic realization for downscaling model output.}

\item{SSE_variable}{variable to use in determining the socioeconomic assignment using the SSE method; can be
Population or PPP-GDP}
}
\description{
Takes in list of model output data from \code{\link{prepare_preprocessed_data}} and returns a list of model-
specific dataframes mapping model scenarios to named nation-level socioeconomic realizations. Default behavior
is to compute these assignments by minimizing sum of squared errors in future GDP ratios by model region and
time period.
}
\details{
Returns a list of dataframes where each model is an element. Data frames assign model's scenarios
to a country-level socioeconomics file.
}
