% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{downscale_flow_variables}
\alias{downscale_flow_variables}
\title{Downscale annual flow variables from model regions to countries}
\usage{
downscale_flow_variables(model_data_list, model_socio_list,
  country_share_list, collapse_list = TRUE, ...)
}
\arguments{
\item{model_data_list}{list of model-specific dataframes with cleaned and pre-processed output}

\item{model_socio_list}{list of model-specific dataframes assigning each submitted scenario to a country-level
socioeconomic scenario}

\item{country_share_list}{list of model-specific lists of downscaling proxy dataframes indicating country-
within-region shares}

\item{collapse_list}{logical (default = TRUE) indicating whether to return output as a single data frame (TRUE) or a
list of model-specific data frames (FALSE).}
}
\description{
Performs a merge of (1) pre-processed and corrected model output (by native model region) from
\code{\link{prepare_preprocessed_data}}, (2) country-within-model-region shares of downscaling proxies from
\code{\link{compute_country_shares}}, and (3) assignment of a nation-level socioeconomic realization for each model
and scenario from \code{\link{assign_socioeconomic_scenario}}. Returns a list of model-specific dataframes with all
quantity (i.e., annual flow) variables reported at the country level.
}
\details{
Takes in three lists; the named elements of each of these lists should be the model names, and they must
  be standardized. All models and scenarios in \code{model_data_list} must be available in \code{model_socio_list},
  though there is no problem if the latter contains additional models and/or scenarios. Similarly, all models and
  socioeconomic realizations from joining \code{model_data_list} and \code{model_socio_list} must be available in
  \code{country_share_list}, though the latter may contain additional nation-level socioeconomic realizations.
  Finally, all reported quantity variables must be assigned to a downscaling proxy in \code{variable_ds_proxy_fn}
  that is also available in \code{country_share_list}.
}
