% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{load_preprocessed_data}
\alias{load_preprocessed_data}
\title{Load the data submitted by the individual models, cleaned and placed into a common folder}
\usage{
load_preprocessed_data(model_data_folder)
}
\arguments{
\item{model_data_folder}{folder where the pre-processed model output data is located}
}
\description{
Load the data submitted by the individual models, cleaned and placed into a common folder
}
\details{
Assumes that the model data are all in the same folder, named with the extension ".csv",
and that no other csv files are in the same folder
}
