% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{prepare_preprocessed_data}
\alias{prepare_preprocessed_data}
\title{Apply corrections to pre-processed data, with options to unlist and interpolate years}
\usage{
prepare_preprocessed_data(data, collapse_list = FALSE,
  interpolate_years = TRUE)
}
\arguments{
\item{data}{list of data frames with model-specific data}

\item{collapse_list}{logical indicating whether to collapse the output data into a single data
frame, or keep it as a list of data frames}

\item{interpolate_years}{logical indicating whether to interpolate years (no extrapolation option)}
}
\description{
Apply corrections to pre-processed data, with options to unlist and interpolate years
}
\details{
Applies model-specific functions, matching the named element of the list to a
corresponding function containing the same string ("correct_" + model name). Also interpolates
years, where expected years are not provided.
}
